fix: Use seed & node data types of consistent size
- Seed: uint64_t
- Fanout: uint16_t
- Level: uint32_t
- Number: uint32_t

+++ daisi/src/manager/general_scenariofile.h
@@ -66,7 +66,7 @@ struct GeneralScenariofile {
   std::string version;
 
   // setup information
-  int random_seed = 0;
+  uint64_t random_seed = 0;
 
   // simulation information
   uint64_t stop_time = 0;
